moving RestAPI to tools project using futures in method calls in the rest server
